refactor: Docs, move files to utils, pass in flowId
